server/requirements.txt
asgiref==3.4.1
Django==3.2.11
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
environ==1.0
Pillow==9.0.0
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
